para search

+++ danger_wash_ckpt.sh
@@ -1,1 +1,19 @@
-find . -name "checkpoint-*" | xargs rm -rf 
+find . -name "checkpoint-*"
+
+read -r -p "Are You Sure? [Y/n] " input
+
+case $input in
+    [yY][eE][sS]|[yY])
+		find . -name "checkpoint-*" | xargs rm -rf 
+        echo "Done"
+		;;
+
+    [nN][oO]|[nN])
+		echo "No"
+       	;;
+
+    *)
+		echo "Invalid input..."
+		exit 1
+		;;
+esac
